backend/requirements.txt
fastapi[standard]
sqlmodel
PyJWT
pwdlib[argon2]
mysql-connector-python
alembic
